Move TestApp test cases
